move generate vectors test
